Add ledger integration tests.
The test tries to install a package, with 1 declared dependency and 1
devDependency; the package and its dependency should be installed, but
not the devDependency.

The compressed packages are built on the files.

As for mypy-based tests, the new tests are triggered based on
a sentinel file used by the Makefile.

This commit also fixes the .gitignore file to only ignore the root
Makefile (built from Makefile.in through ./configure), but not Makefile
in subdirectories.